Expanded capabilities for rio shapes.
Closes #150.
